move and show/hide with flags and checks
