+++ dotcom-rendering/src/layouts/LiveLayout.tsx
@@ -38,7 +38,7 @@ import { Island } from '../components/Island';
 import { KeyEventsCarousel } from '../components/KeyEventsCarousel.importable';
 import { Liveness } from '../components/Liveness.importable';
 import { MainMedia } from '../components/MainMedia';
-import { StickyLiveblogAsk } from '../components/marketing/epics/StickyLiveblogAsk.importable';
+import { StickyLiveblogAsk } from '../components/StickyLiveblogAsk.importable';
 import { Masthead } from '../components/Masthead';
 import { MostViewedFooterData } from '../components/MostViewedFooterData.importable';
 import { MostViewedFooterLayout } from '../components/MostViewedFooterLayout';
@@ -57,10 +57,7 @@ import {
 	TopicFilterBank,
 } from '../components/TopicFilterBank';
 import { canRenderAds } from '../lib/canRenderAds';
-import {
-	getContributionsServiceUrl,
-	// shouldHideSupportMessaging,
-} from '../lib/contributions';
+import { getContributionsServiceUrl } from '../lib/contributions';
 import { decideTrail } from '../lib/decideTrail';
 import { getZIndex } from '../lib/getZIndex';
 import type { NavType } from '../model/extract-nav';
@@ -955,21 +952,22 @@ export const LiveLayout = (props: WebProps | AppsProps) => {
 												}
 											/>
 										</div>
-										{!article.shouldHideReaderRevenue &&
-											// !shouldHideSupportMessaging() &&
-											article.config.switches
-												.brazeTaylorReport !== true && (
-												<Island
-													priority="feature"
-													defer={{ until: 'visible' }}
-												>
-													<StickyLiveblogAsk
-														referrerUrl={
-															article.webURL
-														}
-													/>
-												</Island>
-											)}
+										<Island
+											priority="feature"
+											defer={{ until: 'visible' }}
+										>
+											<StickyLiveblogAsk
+												referrerUrl={article.webURL}
+												shouldHideReaderRevenueOnArticle={
+													article.shouldHideReaderRevenue
+												}
+												isEnabled={
+													article.config.switches
+														.brazeTaylorReport !==
+													true
+												}
+											/>
+										</Island>
 									</Hide>
 								)}
 								{/* Match stats */}
